azure_functions/requirements.txt
# Azure Functions SDK
azure-functions>=1.18.0

# Database
sqlalchemy>=2.0.0
pymssql>=2.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# ML
scikit-learn>=1.3.0
xgboost>=1.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
tabulate>=0.9.0
